Big refactor. Make interpreter use a flat map uuid->node.
This fixed a bunch of infinite recursion errors.
The localStack is now a stack of maps from identifier to uuid, where
the last element on the stack is the current scope of an executing
lambda. When we define a lambda, we attach that last element of the
stack to the ast node (which we call the scope of the lambda). We'll
then merge that scope with the arguments of the lambda when executing it
and push it onto the localStack. This allows us to keep track of the
height of the stack, and throw when it gets too big.
We then store the AST nodes inside a global map uuid->node.